feat: check for authorized user

+++ server/api/car/listings/[listingId].delete.js
@@ -1,8 +1,15 @@
+import { serverSupabaseUser } from "#supabase/server";
 import { PrismaClient } from "@prisma/client";
 
 const prisma = new PrismaClient();
 
 export default defineEventHandler(async (event) => {
+  const user = await serverSupabaseUser(event);
+
+  if (!user.id) {
+    throw createError({ statusCode: 401, statusMessage: "Not Authorized!" });
+  }
+
   const { listingId } = event.context.params;
 
   return prisma.listing.delete({
